requirements.txt
tensorflow>=2.9.0
numpy
opencv-python
scikit-learn
pillow
